Add device_id for telemetry
